Fix potential overwrite of pass-code store
Previously, if `cmd_show` fails, e.g. because gpg fails, pass-code continues with
an empty `codec`. This is not a problem when doing an `ls` but an
`insert` will overwrite the `.passcode` file with just the newly inserted
item.

This fix checks for the success of `cmd_show` and aborts on failure.